// apps/react-frontend/src/api/client/models/sync-conflict-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Eazy Recycling Backend Service
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface SyncConflictDto
 */
export interface SyncConflictDto {
    /**
     * 
     * @type {string}
     * @memberof SyncConflictDto
     */
    'id': string;
    /**
     * 
     * @type {string}
     * @memberof SyncConflictDto
     */
    'companyId': string;
    /**
     * 
     * @type {string}
     * @memberof SyncConflictDto
     */
    'externalId'?: string;
    /**
     * 
     * @type {string}
     * @memberof SyncConflictDto
     */
    'exactGuid'?: string;
    /**
     * 
     * @type {string}
     * @memberof SyncConflictDto
     */
    'syncStatus': string;
    /**
     * 
     * @type {{ [key: string]: any; }}
     * @memberof SyncConflictDto
     */
    'conflictDetails'?: { [key: string]: any; };
    /**
     * 
     * @type {string}
     * @memberof SyncConflictDto
     */
    'syncedFromSourceAt': string;
}